+++ packages/ricos-common/web/src/biCallbacks/getBiCallbacks.ts
@@ -1,0 +1,6 @@
+import { BICallbacks } from 'wix-rich-content-common';
+
+export function getBiCallback<T extends keyof BICallbacks>(key: T): BICallbacks[T] | undefined {
+  const { children, _rcProps } = this.props;
+  return children?.props.helpers?.[key] || _rcProps?.helpers?.[key];
+}