+++ packages/ricos-editor/web/src/toolbars/utils/toolbarsConfig.ts
@@ -23,6 +23,6 @@ export const toolbarSettingsFromConfig = ({
     ? get(toolbarConfig?.getButtons?.() as TextButtons, deviceName, [])
     : null;
   const getTextPluginButtons = get(toolbarConfig?.getTextPluginButtons?.(), deviceName, []);
-  const buttonsMapper = get(toolbarConfig?.buttonsMapper?.(), deviceName, []);
-  return { shouldCreate, getButtons, getTextPluginButtons, buttonsMapper };
+  const buttonsOverrides = get(toolbarConfig?.buttonsOverrides?.(), deviceName, []);
+  return { shouldCreate, getButtons, getTextPluginButtons, buttonsOverrides };
 };