+++ packages/tiptap-editor/web/src/commands/deleteNode.ts
@@ -12,8 +12,8 @@ declare module '@tiptap/core' {
   }
 }
 
-export const deleteNode: RawCommands['deleteNode'] = id => ({ view, tr, dispatch }) => {
-  const nodeWithPos = findChildren(view.state.doc, node => node.attrs.id === id);
+export const deleteNode: RawCommands['deleteNode'] = id => ({ tr, dispatch }) => {
+  const nodeWithPos = findChildren(tr.doc, node => node.attrs.id === id);
   const { pos, node } = nodeWithPos?.[0] || {};
   if (pos !== undefined && dispatch) {
     tr.delete(pos, pos + node.nodeSize);